Move edb.server.cluster to edb.testbase.cluster
It is only used by testing and tool flows, and I always forget that.
