+++ edb/tools/gen_test_dumps.py
@@ -31,7 +31,7 @@ import unittest
 import click
 
 from edb import buildmeta
-from edb.server import cluster as edgedb_cluster
+from edb.testbase import cluster as edgedb_cluster
 from edb.testbase import server as tb
 from edb.tools.edb import edbcommands
