+++ edb/tools/inittestdb.py
@@ -29,7 +29,7 @@ import unittest
 import click
 
 from edb.common import devmode
-from edb.server import cluster as edgedb_cluster
+from edb.testbase import cluster as edgedb_cluster
 from edb.testbase import server as tb
 from edb.tools.edb import edbcommands
